requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
idna==2.7
kombu==4.2.1
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.11.0
urllib3==1.24.1
vine==1.1.4
